+++ packages/sui-compiler-config/src/index.js
@@ -10,9 +10,10 @@ const DEFAULT_BROWSER_TARGETS = {
   ios: '14.5'
 }
 
-export const getSWCConfig = ({isModern, isTypeScript}) => {
+const getSWCConfig = ({isModern = false, isTypeScript = false, compileToCJS = false}) => {
   const targets = isModern ? DEFAULT_BROWSER_TARGETS : DEFAULT_LEGACY_BROWSER_TARGETS
   const syntaxOptions = isTypeScript ? {syntax: 'typescript', tsx: true} : {syntax: 'ecmascript', jsx: true}
+  const moduleOptions = compileToCJS ? {module: {type: 'commonjs'}} : {}
 
   return {
     minify: true,
@@ -39,6 +40,7 @@ export const getSWCConfig = ({isModern, isTypeScript}) => {
       loose: true,
       externalHelpers: true
     },
+    ...moduleOptions,
     env: {
       targets,
       dynamicImport: true,
@@ -48,3 +50,5 @@ export const getSWCConfig = ({isModern, isTypeScript}) => {
     }
   }
 }
+
+module.exports.getSWCConfig = getSWCConfig