+++ packages/sui-lint/eslintrc.js
@@ -211,14 +211,21 @@ module.exports = {
     // },
     {
       files: ['**/*.+(ts|tsx)'],
-      extends: ['standard-with-typescript'],
+      extends: ['standard-with-typescript', 'standard-react', 'prettier'],
+      parser: '@typescript-eslint/parser',
       parserOptions: {
         project: './tsconfig.json'
       },
       rules: {
+        'import/extensions': RULES.OFF,
         'no-return-await': RULES.OFF,
-        'prettier/prettier': RULES.OFF,
-        'react/react-in-jsx-scope': RULES.OFF
+        'prettier/prettier': [RULES.ERROR, prettierOptions],
+        'react/react-in-jsx-scope': RULES.OFF,
+        'react/no-unused-prop-types': RULES.OFF,
+        '@typescript-eslint/explicit-function-return-type': [RULES.OFF, {allowTypedFunctionExpressions: false}],
+        'chai-friendly/no-unused-expressions': RULES.ERROR,
+        '@typescript-eslint/no-unused-expressions': RULES.OFF,
+        '@typescript-eslint/return-await': RULES.OFF
       }
     },
     {