+++ packages/sui-react-head/src/server.ts
@@ -1,8 +1,8 @@
-import { renderToString } from 'react-dom/server'
+import {renderToString} from 'react-dom/server'
 
-import { BODY_ATTRIBUTES_KEY } from './Body'
-import { HTML_ATTRIBUTES_KEY } from './Html'
-import { ComponentTag } from './types'
+import {BODY_ATTRIBUTES_KEY} from './Body'
+import {HTML_ATTRIBUTES_KEY} from './Html'
+import {type ComponentTag} from './types'
 
 interface ExtractPropsFromConfig {
   withKey: string
@@ -12,13 +12,13 @@ interface ExtractPropsFromConfig {
  * Extract props from a list of tags a specific tag by using a key
  * and then discard this key before returning the key
  */
-const extractPropsFrom = (tags: ComponentTag[], { withKey }: ExtractPropsFromConfig): undefined | { [x: string]: any} => {
+const extractPropsFrom = (tags: ComponentTag[], {withKey}: ExtractPropsFromConfig): undefined | Record<string, any> => {
   // search the tag using the key and default to an empty object for simplicity
-  const tag: ComponentTag | undefined = tags.find(({ props }) => props.name === withKey)
+  const tag: ComponentTag | undefined = tags.find(({props}) => props.name === withKey)
 
   if (tag != null) {
     // discard the key used to search the tag
-    const { name, ...restOfTag } = tag.props
+    const {name, ...restOfTag} = tag.props
     // return only the desired info for the tag
     return restOfTag
   }
@@ -27,7 +27,7 @@ const extractPropsFrom = (tags: ComponentTag[], { withKey }: ExtractPropsFromCon
 /**
  * Transform the object from the head to a string to be used in the server
  */
-const transformToString = (headObject: { [key: string]: string} = {}): string =>
+const transformToString = (headObject: Record<string, string> = {}): string =>
   Object.entries(headObject)
     .map(([key, value]) => `${key}="${value}"`)
     .join(' ')
@@ -35,7 +35,11 @@ const transformToString = (headObject: { [key: string]: string} = {}): string =>
 /**
  * Render the tags for the head
  */
-export function renderHeadTagsToString (headTags: any[]): { headString: string, bodyAttributes: string, htmlAttributes: string} {
+export function renderHeadTagsToString(headTags: ComponentTag[]): {
+  headString: string
+  bodyAttributes: string
+  htmlAttributes: string
+} {
   const bodyAttributesProps = extractPropsFrom(headTags, {
     withKey: BODY_ATTRIBUTES_KEY
   })
@@ -44,8 +48,7 @@ export function renderHeadTagsToString (headTags: any[]): { headString: string, 
   })
 
   const headTagsToRender = headTags.filter(
-    ({ props }) =>
-      props.name !== BODY_ATTRIBUTES_KEY && props.name !== HTML_ATTRIBUTES_KEY
+    ({props}) => props.name !== BODY_ATTRIBUTES_KEY && props.name !== HTML_ATTRIBUTES_KEY
   )
 
   return {