+++ packages/sui-studio/src/components/utils.js
@@ -1,3 +1,4 @@
+/* globals __SUI_STUDIO_CONFIG__ __SUI_STUDIO_COMPONENTS_LIST__ */
 import {forwardRef} from 'react'
 
 import hoistNonReactStatics from 'hoist-non-react-statics'
@@ -8,27 +9,11 @@ const DEFAULT_LOGO =
   "<svg viewBox='0 0 470 340.6'> <path fill='#e30613' d='M129.6 170.3c0-58.6 29.6-110.2 74.6-140.8L155.5 1.8c-4.2-2.4-9.3-2.3-13.5.1L6.7 81.2C2.5 83.6 0 88.1 0 92.9v156.9c0 4.8 2.6 9.2 6.8 11.6l137.4 77.5c4.2 2.3 9.4 2.3 13.6-.2l46.7-27.4c-45.2-30.6-74.9-82.3-74.9-141z' /> <path fill='#009fe3' d='M299.8.1c-35.4 0-68.3 10.8-95.6 29.4l87.7 49.8c4.2 2.4 6.8 6.8 6.8 11.6l1.1 156.8c0 4.8-2.5 9.3-6.7 11.7l-88.6 51.9c27.2 18.4 60 29.2 95.3 29.2 94 0 170.2-76.2 170.2-170.2S393.8.1 299.8.1z' /> <path fill='#000411' d='M299.8 247.7l-1.1-156.8c0-4.8-2.6-9.2-6.8-11.6l-87.7-49.8c-45 30.6-74.6 82.3-74.6 140.8 0 58.7 29.7 110.4 74.9 141l88.6-51.9c4.2-2.4 6.7-6.9 6.7-11.7z' /></svg>"
 
 export const getComponentsList = () => {
-  const listOfComponents = preval`
-    const fg = require('fast-glob')
-    const path = require('path')
-
-    const folders = fg.sync('components/*/*', { deep: 3, onlyDirectories: true})
-    const components = folders.map(folder => {
-      const [,category, component] = folder.split(path.sep)
-      return {category, component}
-    })
-    module.exports = components
-  `
-  return listOfComponents
+  return __SUI_STUDIO_COMPONENTS_LIST__
 }
 
 export const getSuiStudioConfig = () => {
-  const config = preval`
-    const path = require('path')
-    const {config = {}} = require(path.join(process.cwd(), 'package.json'))
-    module.exports = config['sui-studio'] || {}
-  `
-  return config
+  return __SUI_STUDIO_CONFIG__
 }
 
 export const getStudioName = () => {