+++ packages/sui-studio/workbench/src/components/Root/index.js
@@ -2,12 +2,12 @@ import {useState} from 'react'
 
 import PropTypes from 'prop-types'
 
-import Header from '../Header/index.js'
-import Select from '../Select/index.js'
-import Test from '../Suite/index.js'
+import Header from '../Header/index'
+import Select from '../Select/index'
+import Test from '../Suite/index'
 
-const importComponent = () => import('component/index.js')
-const importTest = () => import('test/index.test.js')
+const importComponent = () => import('component/index')
+const importTest = () => import('test/index.test')
 
 const getFromStorage = (key, defaultValue) => window.sessionStorage[key] || defaultValue